# packages/server/main.go
package main

import (
	"fmt"
	"log/slog"
	"os"
	"os/signal"
	"strconv"
	"strings"
	"syscall"
	"time"

	_ "embed"

	"github.com/gofiber/fiber/v2"
	"github.com/gofiber/fiber/v2/middleware/cors"
	"github.com/gofiber/fiber/v2/middleware/logger"
	"github.com/gofiber/fiber/v2/middleware/recover"
	"vaporrmm/server/internal/auth"
	"vaporrmm/server/internal/db"
	"vaporrmm/server/internal/events"
	"vaporrmm/server/internal/handlers"
	"vaporrmm/server/internal/metrics"
	"vaporrmm/server/internal/redis"
	"vaporrmm/server/internal/utils"
)

//go:embed openapi.json
var openAPISpec []byte

var (
	buildVersion = "dev"
)

const (
	defaultServerPort       = 8080
	defaultAgentWSPort      = 47991
	defaultOfflineThreshold = 120
	defaultBodyLimit        = 4 * 1024 * 1024
	defaultReadTimeout      = 30 * time.Second
	defaultWriteTimeout     = 30 * time.Second
	defaultIdleTimeout      = 120 * time.Second
	defaultCookieMaxAge     = 86400
	defaultAgentPort        = 47991
	defaultSunshinePort     = 47990
	maxDevicesLimit         = 500
	maxCommandLimit         = 200
	maxAuditLimit           = 1000
	metricsRetentionDefault = 86400
	defaultTickerInterval   = 60 * time.Second
	defaultHSTSMaxAge       = 31536000
)

func main() {
	// Load configuration from environment / Docker secrets
	auth.JWTSecret = utils.ReadSecret("JWT_SECRET", "JWT_SECRET_FILE")
	if auth.JWTSecret == "" {
		auth.JWTSecret = utils.GenerateSecureKey()
		slog.Info("Warning: JWT_SECRET not set, using generated key (will not persist across restarts)")
	}

	utils.ServerPort = defaultServerPort
	if p := os.Getenv("SERVER_PORT"); p != "" {
		if parsed, err := strconv.Atoi(p); err == nil && parsed > 0 {
			utils.ServerPort = parsed
		}
	}

	utils.AgentWSPort = defaultAgentWSPort
	if p := os.Getenv("AGENT_WS_PORT"); p != "" {
		if parsed, err := strconv.Atoi(p); err == nil && parsed > 0 {
			utils.AgentWSPort = parsed
		}
	}

	moonlightWebURL := os.Getenv("MOONLIGHT_WEB_URL")

	if err := db.Init(); err != nil {
		slog.Error("failed to initialize database", "error", err)
		os.Exit(1)
	}
	defer db.DB.Close()

	redis.Init()
	defer redis.Close()

	events.StartWSRedisSubscriber()

	auth.LoadAgentTokens()
	auth.CreateDefaultAdmin()

	app := fiber.New(fiber.Config{
		BodyLimit:             defaultBodyLimit,
		ReadTimeout:           defaultReadTimeout,
		WriteTimeout:          defaultWriteTimeout,
		IdleTimeout:           defaultIdleTimeout,
		DisableStartupMessage: false,
	})

	app.Use(recover.New())
	app.Use(logger.New(logger.Config{
		Format: "${time} ${status} ${method} ${path} ${latency}\n",
	}))

	// Request ID / trace context middleware
	app.Use(func(c *fiber.Ctx) error {
		traceID := c.Get("X-Trace-ID")
		if traceID == "" {
			traceID = utils.GenerateSecureKey()[:16]
		}
		c.Set("X-Trace-ID", traceID)
		c.Locals("trace_id", traceID)
		return c.Next()
	})

	// Security headers middleware
	app.Use(func(c *fiber.Ctx) error {
		c.Set("X-Content-Type-Options", "nosniff")
		c.Set("X-Frame-Options", "DENY")
		c.Set("Referrer-Policy", "strict-origin-when-cross-origin")
		c.Set("Content-Security-Policy", "default-src 'self'")
		return c.Next()
	})

	// HTTPS redirect + HSTS when TLS_CERT is set
	if os.Getenv("SERVER_CERT") != "" {
		app.Use(func(c *fiber.Ctx) error {
			if c.Protocol() == "http" {
				return c.Redirect("https://"+c.Hostname()+c.OriginalURL(), fiber.StatusMovedPermanently)
			}
			c.Set("Strict-Transport-Security", fmt.Sprintf("max-age=%d; includeSubDomains; preload", defaultHSTSMaxAge))
			return c.Next()
		})
	}

	// CORS
	corsOrigins := os.Getenv("CORS_ORIGINS")
	if corsOrigins == "" {
		corsOrigins = "http://localhost:3000"
	}
	app.Use(cors.New(cors.Config{
		AllowOrigins:     corsOrigins,
		AllowMethods:     "GET,POST,PUT,DELETE,OPTIONS",
		AllowHeaders:     "Origin,Content-Type,Authorization,X-CSRF-Token",
		AllowCredentials: true,
	}))

	// Prometheus metrics middleware
	app.Use(func(c *fiber.Ctx) error {
		start := time.Now()
		err := c.Next()
		duration := time.Since(start).Seconds()
		status := strconv.Itoa(c.Response().StatusCode())
		path := c.Route().Path
		if path == "" {
			path = c.Path()
		}
		metrics.HTTPRequestsTotal.WithLabelValues(c.Method(), path, status).Inc()
		metrics.HTTPRequestDuration.WithLabelValues(c.Method(), path).Observe(duration)
		return err
	})

	cfg := handlers.Config{
		BuildVersion:            buildVersion,
		DefaultOfflineThreshold: defaultOfflineThreshold,
		DefaultAgentWSPort:      defaultAgentWSPort,
		DefaultSunshinePort:     defaultSunshinePort,
		DefaultCookieMaxAge:     defaultCookieMaxAge,
		MaxDevicesLimit:         maxDevicesLimit,
		MaxCommandLimit:         maxCommandLimit,
		MaxAuditLimit:           maxAuditLimit,
		MoonlightWebURL:         moonlightWebURL,
	}

	// System routes
	handlers.RegisterSystemRoutes(app, cfg, openAPISpec)

	// Agent routes
	handlers.RegisterAgentRoutes(app, cfg)

	// Public API group
	publicAPI := app.Group("/api", auth.RateLimiter(60, time.Minute))

	// API v1 routes
	api := app.Group("/api/v1", auth.AuthMiddleware(), auth.CSRFMiddleware())

	// Backward compatibility: redirect legacy /api/* paths to /api/v1/*
	// Uses 308 (Permanent Redirect) to preserve HTTP method (POST, PUT, etc.)
	// Public endpoints (/api/auth/*, /api/branding/*) are excluded because they
	// must remain accessible without CSRF / auth middleware.
	app.Use(func(c *fiber.Ctx) error {
		path := c.Path()
		if strings.HasPrefix(path, "/api/") && !strings.HasPrefix(path, "/api/v1/") &&
			path != "/api/version" && path != "/api/openapi.json" &&
			!strings.HasPrefix(path, "/api/auth/") &&
			!strings.HasPrefix(path, "/api/branding/") {
			return c.Redirect("/api/v1"+strings.TrimPrefix(path, "/api"), fiber.StatusPermanentRedirect)
		}
		return c.Next()
	})

	// Auth routes
	handlers.RegisterAuthRoutes(publicAPI, api, cfg)

	// Branding routes
	handlers.RegisterBrandingRoutes(app, api)

	// Device routes
	devices := api.Group("/devices")
	handlers.RegisterDeviceRoutes(api, devices, cfg)

	// Dashboard routes
	handlers.RegisterDashboardRoutes(api, cfg)

	// Script routes
	handlers.RegisterScriptRoutes(api, cfg)

	// Compliance routes
	handlers.RegisterComplianceRoutes(api, cfg)

	// Webhook routes
	handlers.RegisterWebhookRoutes(api)

	// Audit routes
	handlers.RegisterAuditRoutes(api, cfg)

	// Alert routes
	handlers.RegisterAlertRoutes(api)

	// Admin routes
	handlers.RegisterAdminRoutes(api)

	// Ticket routes
	handlers.RegisterTicketRoutes(api, cfg)

	// ============================================================
	// Background goroutines
	// ============================================================
	offlineSec := defaultOfflineThreshold
	if o := os.Getenv("OFFLINE_THRESHOLD_SECONDS"); o != "" {
		if parsed, err := strconv.Atoi(o); err == nil && parsed > 0 {
			offlineSec = parsed
		}
	}
	slog.Info("offline threshold configured", "seconds", offlineSec)

	offlineDone := make(chan struct{})
	go func() {
		ticker := time.NewTicker(defaultTickerInterval)
		defer ticker.Stop()
		for {
			select {
			case <-ticker.C:
				threshold := time.Now().Unix() - int64(offlineSec)
				res, err := db.DB.Exec(`UPDATE devices SET status = 'offline' WHERE last_seen < ? AND status != 'offline'`, threshold)
				if err != nil {
					slog.Warn("offline detection query failed", "error", err)
					continue
				}
				if n, _ := res.RowsAffected(); n > 0 {
					slog.Info("marked devices offline", "count", n)
					rows, _ := db.DB.Query(`SELECT id, hostname FROM devices WHERE last_seen < ? AND status = 'offline' LIMIT ?`, threshold, n)
					if rows != nil {
						for rows.Next() {
							var did, hostname string
							_ = rows.Scan(&did, &hostname)
							events.TriggerWebhooks("device.offline", map[string]interface{}{"device_id": did, "hostname": hostname, "timestamp": time.Now().Unix()})
							events.WSBroadcastMessage(map[string]interface{}{"type": "device.offline", "device_id": did, "hostname": hostname, "timestamp": time.Now().Unix()})
						}
						rows.Close()
					}
				}
			case <-offlineDone:
				return
			}
		}
	}()

	go func() {
		ticker := time.NewTicker(defaultTickerInterval)
		defer ticker.Stop()
		for {
			select {
			case <-ticker.C:
				var total, online int64
				_ = db.DB.QueryRow(`SELECT COUNT(*) FROM devices`).Scan(&total)
				_ = db.DB.QueryRow(`SELECT COUNT(*) FROM devices WHERE status = 'online'`).Scan(&online)
				metrics.RegisteredDevicesGauge.Set(float64(total))
				metrics.ActiveDevicesGauge.Set(float64(online))
				if db.DB != nil && db.DB.DB != nil {
					stats := db.DB.DB.Stats()
					metrics.DBOpenConnsGauge.Set(float64(stats.OpenConnections))
					metrics.DBInUseConnsGauge.Set(float64(stats.InUse))
					metrics.DBIdleConnsGauge.Set(float64(stats.Idle))
				}
			case <-offlineDone:
				return
			}
		}
	}()

	// Graceful shutdown on SIGINT/SIGTERM
	quit := make(chan os.Signal, 1)
	signal.Notify(quit, os.Interrupt, syscall.SIGTERM)
	go func() {
		<-quit
		slog.Info("shutting down server...")
		close(offlineDone)
		if err := app.Shutdown(); err != nil {
			slog.Error("server shutdown error", "error", err)
		}
	}()

	// Start server
	slog.Info("starting server", "port", utils.ServerPort)
	if err := app.Listen(fmt.Sprintf(":%d", utils.ServerPort)); err != nil {
		slog.Error("failed to start server", "error", err)
	}
}