# packages/server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vapor RMM API",
    "description": "API for managing devices, commands, and monitoring through the Vapor RMM platform.",
    "version": "1.0.0",
    "contact": {
      "name": "Vapor RMM Support",
      "url": "https://github.com/vapor-rmm/vapor-rmm"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Current server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/branding": {
      "get": {
        "summary": "Get branding configuration",
        "tags": ["Branding"],
        "responses": {
          "200": {
            "description": "Branding configuration",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "app_name": { "type": "string" },
                    "icon_url": { "type": "string" },
                    "company_name": { "type": "string" },
                    "primary_color": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update branding configuration",
        "tags": ["Branding"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "app_name": { "type": "string" },
                  "icon_url": { "type": "string" },
                  "company_name": { "type": "string" },
                  "primary_color": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Branding updated" }
        }
      }
    },
    "/devices": {
      "get": {
        "summary": "List all devices",
        "tags": ["Devices"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "os", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "sort_by", "in": "query", "schema": { "type": "string" } },
          { "name": "sort_order", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } },
          { "name": "offset", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "List of devices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "devices": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Device" }
                    },
                    "total": { "type": "integer" },
                    "has_more": { "type": "boolean" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new device",
        "tags": ["Devices"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DeviceInput" }
            }
          }
        },
        "responses": {
          "201": { "description": "Device created" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Device": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "hostname": { "type": "string" },
          "ip_address": { "type": "string" },
          "mac_address": { "type": "string" },
          "os_name": { "type": "string" },
          "os_version": { "type": "string" },
          "agent_version": { "type": "string" },
          "status": { "type": "string", "enum": ["online", "offline", "maintenance"] },
          "last_seen": { "type": "integer" },
          "created_at": { "type": "integer" },
          "cpu": { "type": "string" },
          "memory": { "type": "integer" },
          "disk_size": { "type": "integer" },
          "tags": { "type": "array", "items": { "type": "string" } }
        }
      },
      "DeviceInput": {
        "type": "object",
        "required": ["name", "hostname"],
        "properties": {
          "name": { "type": "string" },
          "hostname": { "type": "string" },
          "ip_address": { "type": "string" },
          "mac_address": { "type": "string" },
          "os_name": { "type": "string" },
          "os_version": { "type": "string" }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}